Keep the pot and stacks under PioSOLVER's 16-bit chip limit
PioViewer names the real cause of the failed solves:

  set_eff_stack Both stack sizes and the starting pot should be below
  65535 (consider dividing stacks/pot by 10 or 100)

65535 is 2^16-1, so this is a hard ceiling and not a size heuristic. A
$735 effective stack at 100 chips per dollar is 73500 and was refused
outright, which is why go and dump_tree then found no tree. It is a
regression from denominating solves in the hand's own money: the same
stack under the old convention was 245bb, or 24500 chips.

pickChipScale now applies that ceiling first and treats it as absolute,
then satisfies wholeness and pot resolution within whatever room is left.
The three pulls do not point the same way, so the previous claim that
they could never conflict was simply wrong: $735 into a $31 pot cannot
have both a 500-chip pot and a stack under 65535, and the ceiling has to
win. The failing hand now picks x10 - a 310-chip pot and a 7350-chip
stack - and a 33% bet still resolves to within a third of a percent.

The ladder also extends below 1, because a hand played for more chips
than Pio can hold has to be divided down rather than multiplied up, which
is exactly what the error message suggests. ChipScale is a double through
the API and a float in the watcher to carry those.

Drops the depth warning added in the previous commit: it fit the evidence
available then but described a correlation, not the cause, and the ceiling
is now enforced rather than warned about. The panel instead warns only in
the case that remains genuinely unsolvable, a hand-edited pot or stack
above the limit.

+++ watcher/watch_adls_and_run_pio_headless.py
@@ -839,7 +839,7 @@ def process_gametree_json(
     is_icm: Optional[bool] = None
     seat_meta: Optional[list] = None
     hand_bb: Optional[float] = None
-    chip_scale: Optional[int] = None
+    chip_scale: Optional[float] = None
 
     try:
         obj = json.loads(raw)
